requirements.txt
anyio==3.6.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==3.20.3
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
